backend/v2/middleware/request_id.py
from __future__ import annotations

from starlette.middleware.base import BaseHTTPMiddleware, RequestResponseEndpoint
from starlette.requests import Request
from starlette.responses import Response
from uuid import uuid4

from backend.v2.shared.logger import set_request_context


class RequestIDMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request: Request, call_next: RequestResponseEndpoint) -> Response:
        request_id = str(uuid4())
        request.state.request_id = request_id
        set_request_context(request_id)
        response = await call_next(request)
        response.headers["X-Request-ID"] = request_id
        return response
